+++ src/js/core/sticky.js
@@ -99,9 +99,9 @@ export default {
     },
 
     observe: [
-        resize({ target: ({ $el }) => [$el, document.scrollingElement] }),
         viewport(),
-        scroll(),
+        scroll({ target: () => document.scrollingElement }),
+        resize({ target: ({ $el }) => [$el, document.scrollingElement] }),
     ],
 
     events: [