+++ src/js/mixin/slider-nav.js
@@ -9,9 +9,10 @@ import {
     empty,
     isNumeric,
     matches,
+    parent,
     toFloat,
-    toggleClass,
     toNumber,
+    toggleClass,
 } from 'uikit-util';
 import { generateId } from '../api/instance';
 import { keyMap } from '../util/keys';
@@ -205,8 +206,7 @@ export default {
     methods: {
         updateNav() {
             const index = this.getValidIndex();
-            let focus;
-            let focusEl;
+
             for (const el of this.navItems) {
                 const cmd = data(el, this.attrItem);
                 const button = $('a,button', el) || el;
@@ -222,11 +222,9 @@ export default {
                         tabindex: active ? null : -1,
                     });
 
-                    if (active) {
-                        focusEl = button;
+                    if (active && button && matches(parent(el), ':focus-within')) {
+                        button.focus();
                     }
-
-                    focus ||= matches(button, ':focus');
                 } else {
                     toggleClass(
                         el,
@@ -235,10 +233,6 @@ export default {
                             ((cmd === 'previous' && index === 0) ||
                                 (cmd === 'next' && index >= this.maxIndex)),
                     );
-                }
-
-                if (focus && focusEl) {
-                    focusEl.focus();
                 }
             }
         },