+++ CHANGELOG.md
@@ -17,6 +17,12 @@
 - The `closest()` function no longer accepts an array of elements
 - Reset fieldset min-width
 
+## WIP
+
+### Added
+
+- Add `-source` argument to `prefix` command
+
 ## 3.16.24 (August 1, 2023)
 
 ### Fixed