+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Add `-source` argument to `prefix` command
 
+### Fixed
+
+- Fix Toggle in hover mode does not trigger when scrolling in FireFox
+
 ## 3.16.24 (August 1, 2023)
 
 ### Fixed