+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixed
 
 - Fix Toggle in hover mode does not trigger when scrolling in FireFox
+- Fix Parallax component stop parsing
 
 ## 3.16.24 (August 1, 2023)