+++ CHANGELOG.md
@@ -24,6 +24,11 @@
 
 - Add `-source` argument to `prefix` command
 
+### Fixed
+
+- Fix Toggle in hover mode does not trigger when scrolling in FireFox
+- Fix Parallax component stop parsing
+
 ## 3.16.24 (August 1, 2023)
 
 ### Fixed